added new changes

+++ src/main/java/com/example/demo/repository/CarRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Car;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
+public interface CarRepository extends JpaRepository<Car, Long> {  // Use Long here, not Integer
 
     // Custom query to find cars by status
     List<Car> findByStatus(String status);  // This will return all cars with the given status
